pyproject.toml
[project]
name = "data-analysis-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain-groq>=1.1.2",
    "python-dotenv>=1.2.2",
]
